+++ lm-api/tests/test_security.py
@@ -6,20 +6,17 @@ from lm_api.security import IdentityPayload, get_domain_configs
 def test_get_domain_configs__loads_only_base_settings(tweak_settings):
     with tweak_settings(
         ARMASEC_DOMAIN="foo.io",
-        ARMASEC_AUDIENCE="https://bar.dev",
     ):
         domain_configs = get_domain_configs()
 
     assert len(domain_configs) == 1
     first_config = domain_configs.pop()
     assert first_config.domain == "foo.io"
-    assert first_config.audience == "https://bar.dev"
 
 
 def test_get_domain_configs__loads_admin_settings_if_all_are_present(tweak_settings):
     with tweak_settings(
         ARMASEC_DOMAIN="foo.io",
-        ARMASEC_AUDIENCE="https://bar.dev",
         ARMASEC_ADMIN_DOMAIN="admin.io",
     ):
         domain_configs = get_domain_configs()
@@ -27,13 +24,10 @@ def test_get_domain_configs__loads_admin_settings_if_all_are_present(tweak_setti
     assert len(domain_configs) == 1
     first_config = domain_configs.pop()
     assert first_config.domain == "foo.io"
-    assert first_config.audience == "https://bar.dev"
 
     with tweak_settings(
         ARMASEC_DOMAIN="foo.io",
-        ARMASEC_AUDIENCE="https://bar.dev",
         ARMASEC_ADMIN_DOMAIN="admin.io",
-        ARMASEC_ADMIN_AUDIENCE="https://admin.dev",
         ARMASEC_ADMIN_MATCH_KEY="foo",
         ARMASEC_ADMIN_MATCH_VALUE="bar",
     ):
@@ -42,9 +36,7 @@ def test_get_domain_configs__loads_admin_settings_if_all_are_present(tweak_setti
     assert len(domain_configs) == 2
     (first_config, second_config) = domain_configs
     assert first_config.domain == "foo.io"
-    assert first_config.audience == "https://bar.dev"
     assert second_config.domain == "admin.io"
-    assert second_config.audience == "https://admin.dev"
     assert second_config.match_keys == dict(foo="bar")