+++ lm-api/CHANGELOG.md
@@ -10,6 +10,8 @@ This file keeps track of all notable changes to `License Manager API`.
 * Upgrade py-buzz to *^4.1.0*
 * Upgrade Pydantic to *^2.8.2* [[PENG-2280](https://sharing.clickup.com/t/h/c/18022949/PENG-2280/YUSOZKBIF96CZJ0)]
 * Add *pydantic-settings* as a dependency
+* Remove the audience setting [[PENG-2231](https://sharing.clickup.com/t/h/c/18022949/PENG-2231/T3PXA6KD1EH124G)]
+
 
 ## 3.3.0 -- 2024-05-30
 * Add new endpoints to manage the cluster reports from the agent [ASP-4601]